reworked the script function source api after review: own global object, no player unit filter, replay compatible text, default on
